requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.0
matplotlib==3.7.2
jetson-inference
jetson-utils
PyQt5==5.15.9
scikit-learn==1.3.0
tqdm==4.66.1
